requirements.txt
# Web framework and ASGI server
fastapi==0.115.12
uvicorn==0.34.0

# Anthropic SDK — used directly (raw) in Phase 1; no LangChain yet
anthropic==0.49.0

# Pydantic for data validation and typed models
# pydantic is bundled with FastAPI, but pydantic-settings is a separate package
# needed for BaseSettings (reading config from .env)
pydantic==2.11.1
pydantic-settings==2.8.1

# Loads .env file into environment variables at process startup
python-dotenv==1.1.0

# Required by FastAPI to parse multipart/form-data (file uploads)
# Without this, uploading an image will raise a 422 error with no clear explanation
python-multipart==0.0.20

# Frontend
streamlit==1.44.0
Pillow==11.2.1  # image preview in Streamlit UI

# HTTP client used by Streamlit to call the FastAPI backend
requests==2.32.3